halo/requirements.txt
gevent==1.0
flask==0.10.1
peewee==2.2.2
flask-peewee==0.6.5
pymysql==0.6.1
mailer==0.7
itsdangerous==0.21
stripe==1.11.0
blinker==1.2
celery==3.1.8
boto==2.23.0 # required for SQS
raven[flask]==4.0.3
pycrypto==2.6.1 # for android purchases
python-gcm==0.1.5
apns==1.1.2

# for deployment
gunicorn==18.0
setproctitle==1.1.7

# for services
oset==0.1.3
tweepy==2.2
facebook-sdk==0.4.0
python-instagram==0.8.0

# for stats models
python-dateutil==1.5
pytz

# custom libraries
#git+git://github.com/kwellman/flask-playground.git
